+++ src/Kernels/Distance/Diagonal.php
@@ -7,9 +7,9 @@ use Rubix\ML\DataType;
 /**
  * Diagonal
  *
- * The Diagonal (sometimes called *Chebyshev*) distance is a measure that constrains
- * movement to horizontal, vertical, and diagonal movement from a point. An example
- * of a game that uses diagonal movement is a chess board.
+ * The Diagonal (a.k.a. *Chebyshev*) distance is a measure that constrains
+ * movement to horizontal, vertical, and diagonal movement from a point. An
+ * example of a game that uses diagonal movement is a chess board.
  *
  * @category    Machine Learning
  * @package     Rubix/ML